decouple the getDisplayedValue logic for unit testing from useMainValue and add a unit test for it - without call the test.
